:boom: Fix:Validação se localStorage Existe para rodar build

+++ src/services/BalanceFetch.ts
@@ -7,22 +7,23 @@ interface BalanceData {
   valor_inicial: number;
 }
 
-const id = localStorage.getItem("id");
-
 export const getBalancesPerPerson = async () => {
-  try {
-    const response = await axios.get(url + id, {
-      headers: {
-        Authorization: `Bearer ${localStorage.getItem("token")}`,
-      },
-    });
-    if (!response.data) {
-      throw new Error(`Erro na requisição: ${response.status}`);
-    }
-    return response.data;
-  } catch (error: any) {
-    if (axios.isAxiosError(error)) {
-      throw new Error(`Erro ao buscar saldos: ${error.message}`);
+  if (typeof localStorage !== undefined) {
+    const id = localStorage.getItem("id");
+    try {
+      const response = await axios.get(url + id, {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem("token")}`,
+        },
+      });
+      if (!response.data) {
+        throw new Error(`Erro na requisição: ${response.status}`);
+      }
+      return response.data;
+    } catch (error: any) {
+      if (axios.isAxiosError(error)) {
+        throw new Error(`Erro ao buscar saldos: ${error.message}`);
+      }
     }
   }
 };
@@ -32,52 +33,57 @@ export const createBalance = async (
   descricao: string,
   valor_inicial: number
 ): Promise<BalanceData> => {
-  try {
-    const response = await axios.post(
-      url + id,
-      {
-        nome,
-        descricao,
-        valor_inicial,
-      },
-      {
-        headers: {
-          Authorization: `Bearer ${localStorage.getItem("token")}`,
+  if (typeof localStorage !== undefined) {
+    const id = localStorage.getItem("id");
+    try {
+      const response = await axios.post(
+        url + id,
+        {
+          nome,
+          descricao,
+          valor_inicial,
         },
+        {
+          headers: {
+            Authorization: `Bearer ${localStorage.getItem("token")}`,
+          },
+        }
+      );
+
+      if (!response.data) {
+        throw new Error(`Erro na requisição: ${response.status}`);
       }
-    );
 
-    if (!response.data) {
-      throw new Error(`Erro na requisição: ${response.status}`);
+      return response.data;
+    } catch (error: any) {
+      if (axios.isAxiosError(error)) {
+        throw new Error(`Erro ao criar saldo: ${error.message}`);
+      }
     }
-
-    return response.data;
-  } catch (error: any) {
-    if (axios.isAxiosError(error)) {
-      throw new Error(`Erro ao criar saldo: ${error.message}`);
-    }
-
-    throw new Error(`Erro interno no servidor`);
   }
+  throw new Error(`Erro interno no servidor`);
 };
 
 export const deleteBalanceByBalanceId = async (balanceId: string) => {
-  try {
-    const response = await axios.delete(
-      url + "delete/" + id + "/?balanceId=" + balanceId,
-      {
-        headers: {
-          Authorization: `Bearer ${localStorage.getItem("token")}`,
-        },
+  if (typeof localStorage !== undefined) {
+    const id = localStorage.getItem("id");
+    try {
+      const response = await axios.delete(
+        url + "delete/" + id + "/?balanceId=" + balanceId,
+        {
+          headers: {
+            Authorization: `Bearer ${localStorage.getItem("token")}`,
+          },
+        }
+      );
+      if (!response.data) {
+        throw new Error(`Erro na requisição: ${response.status}`);
       }
-    );
-    if (!response.data) {
-      throw new Error(`Erro na requisição: ${response.status}`);
-    }
-    return response.data;
-  } catch (error: any) {
-    if (axios.isAxiosError(error)) {
-      throw new Error(`Erro ao deletar saldo: ${error.message}`);
+      return response.data;
+    } catch (error: any) {
+      if (axios.isAxiosError(error)) {
+        throw new Error(`Erro ao deletar saldo: ${error.message}`);
+      }
     }
   }
 };
@@ -88,27 +94,30 @@ export const updateBalanceById = async (
   nome: string,
   descricao: string
 ): Promise<BalanceData> => {
-  try {
-    const response = await axios.patch(
-      url + "update/" + id + "/?balanceId=" + balanceId,
-      {
-        valor_inicial,
-        nome,
-        descricao,
-      },
-      {
-        headers: {
-          Authorization: `Bearer ${localStorage.getItem("token")}`,
+  if (typeof localStorage !== undefined) {
+    const id = localStorage.getItem("id");
+    try {
+      const response = await axios.patch(
+        url + "update/" + id + "/?balanceId=" + balanceId,
+        {
+          valor_inicial,
+          nome,
+          descricao,
         },
+        {
+          headers: {
+            Authorization: `Bearer ${localStorage.getItem("token")}`,
+          },
+        }
+      );
+      if (!response.data) {
+        throw new Error(`Erro na requisição: ${response.status}`);
       }
-    );
-    if (!response.data) {
-      throw new Error(`Erro na requisição: ${response.status}`);
-    }
-    return response.data;
-  } catch (error: any) {
-    if (axios.isAxiosError(error)) {
-      throw new Error(`Erro ao atualizar saldo: ${error.message}`);
+      return response.data;
+    } catch (error: any) {
+      if (axios.isAxiosError(error)) {
+        throw new Error(`Erro ao atualizar saldo: ${error.message}`);
+      }
     }
   }
   throw new Error(`Erro interno no servidor`);
